Shooting working, need collision detection

+++ src/TankTwo.java
@@ -116,7 +116,7 @@ public class TankTwo {
         if(e.getKeyCode() == KeyEvent.VK_S)
             movebackward = true;
          if(e.getKeyCode() == KeyEvent.VK_Q)
-            game.m.add(new Missile(this.game, (int)this.x, (int)this.y, Math.sin(a), Math.cos(a) ));
+            game.m.add(new Missile(this.game, this.x, this.y, Math.sin(a), Math.cos(a) ));
     }
 }
